requirements.txt
pyyaml==3.11
psutil==3.1.1
nltk==3.0.4
simplejson==3.8.0
PyStemmer==1.3.0
numpy==1.9.2
scipy==0.16.0
scikit-learn==0.16.1
pymongo==2
